+++ src/Services/discount/discount.utils.ts
@@ -1,0 +1,8 @@
+import CONSTANTS from "@/resources/resource.contants";
+
+export function buildHeadersWithBasicAuth(): { Authorization: string, 'Content-Type': string } {
+    return {
+        Authorization: process.env.ffs_auth as string,
+        'Content-Type': CONSTANTS.GLOBAL.JSON
+    }
+}
