--- src/Services/historical-price/models/FinancialData.ts
@@ -1,17 +1,0 @@
-export interface FinancialDataListWrapper {
-    quoteSummary: {
-        result: FinancialDataWrapper[];
-    },
-    error: string
-}
-
-export interface FinancialDataWrapper {
-    financialData: FinancialData;
-}
-
-export interface FinancialData {
-    currentPrice: {
-        raw: number,
-        fmt: string
-    }
-}
