--- src/Services/historical-price/models/Frequency.ts
@@ -1,6 +1,0 @@
-export enum Frequency {
-    DAILY = '1d',
-    WEEKLY = '1w',
-    MONTHLY = '1m',
-    YEAR = '1v'
-}
