--- src/Services/historical-price/models/HistoricalPriceInput.ts
@@ -1,8 +1,0 @@
-import { Frequency } from "./Frequency";
-
-export default interface HistoricalPriceInput {
-    symbol: string;
-    fromDate: number;
-    toDate: number;
-    frequency: Frequency
-}
