--- src/Services/historical-price/models/PriceData.ts
@@ -1,4 +1,0 @@
-export default interface PriceData {
-    date: Date,
-    close: number
-}
