+++ src/resources/consumers/consumer.typings.ts
@@ -1,0 +1,3 @@
+export default interface Consumer {
+    startPolling(): Promise<void>   
+}
