--- src/utils/interfaces/IConsumer.ts
@@ -1,5 +1,0 @@
-export default interface Consumer {
-
-    startPolling(): Promise<void>
-    
-}
