--- src/utils/serviceUtils.ts
@@ -1,27 +1,0 @@
-import CONSTANTS from "../resources/resource.contants";
-
-export function buildHeadersWithBasicAuth(): { Authorization: string, 'Content-Type': string } {
-    return {
-        Authorization: process.env.ffs_auth as string,
-        'Content-Type': CONSTANTS.GLOBAL.JSON
-    }
-}
-
-export function buildHistoricalPriceHeaders(): any {
-    return {
-        Accept: 'text/html,application/xhtml+xml,application/xml;q=0.9,image/avif,image/webp,image/apng,*/*;q=0.8,application/signed-exchange;v=b3;q=0.7',
-        'Accept-Encoding': 'gzip, deflate, br',
-        'Accept-Language': 'en-US,en;q=0.9',
-        Connection: 'keep-alive',
-        Host: 'query1.finance.yahoo.com',
-        'Sec-Fetch-Dest': 'document',
-        'Sec-Fetch-Mode': 'navigate',
-        'Sec-Fetch-Site': 'none',
-        'Sec-Fetch-User': '?1',
-        'Upgrade-Insecure-Requests': 1,
-        'User-Agent': 'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/114.0.0.0 Safari/537.36',
-        'sec-ch-ua': '"Not.A/Brand";v="8", "Chromium";v="114", "Google Chrome";v="114"',
-        'sec-ch-ua-mobile': '?0',
-        'sec-ch-ua-platform': "Windows"
-    }
-}
